Tactical Asserts -> Adds the tactical assert library
This is version 1 of the library which works much better than the old design and can be sub classed as a part of any model.

I'd like to eventually add a method to Model which enables tactical asserts after a model has already been instantiated, or during instantiations through an argument flag... or maybe just always use it, hard to say.