Grid Organism -> Renames unused (but required) argument to `_`
`..._add_components_to_segments(self, payload)` is now `..._add_components_to_segments(self, _)`

The `payload` argument position is required as it is a part of a common pattern but for this particular function within the pattern it isn't used.  `_` as an argument name is a way of indicating the position is required but the value is ignored/unused.