text_input.py -> removes use of Keys.NULL before sending keys
In the newest versions of Selenium running on Linux, Keys.NULL sends a
visible character to the text input box.  This was not true until the
last month; I don't know how it changed and no information online seems
to point to the cause.  It might be a new way the ChromeDriver handles
Keys.NULL, it's difficult to say.

Either way, I'm not sure Keys.NULL is required here anyway, it was added
as a precautionary thing a while ago.  If there are new problems with
the keys not being sent in all browsers, I'll look for a new solution.